Revert "feat(task_sdk): make airflow.sdk.definitions.decoratos a package"
This reverts commit 324efc079ec1c5c2618bb19d48d188a1363f3931.
