+++ airflow/example_dags/example_asset_decorator.py
@@ -20,7 +20,7 @@ import pendulum
 
 from airflow.decorators import dag, task
 from airflow.sdk.definitions.asset import Asset
-from airflow.sdk.definitions.decorators.asset import asset
+from airflow.sdk.definitions.asset.decorators import asset
 
 
 @asset(uri="s3://bucket/asset1_producer", schedule=None)
